mv funs to ArHelpers
